correct to float

+++ src/data_types.hpp
@@ -23,7 +23,7 @@ public:
     std::vector<std::vector<int> > genTimeInfec;
     // std::vector<std::vector<int> > genDegreeVec;
 
-    std::vector<std::vector<int> > genMoved;
+    std::vector<std::vector<float> > genMoved;
     std::vector<int> genInfected;
     std::vector<int> gens;
     std::vector<float> pSrcInfect;
